requirements.txt
pathway
litellm
streamlit
python-dotenv
pandas